Added an authorisation hook, context, and modal.
To display the sign-in page, the sign in page has been made into a modal. This was done due to rendering issues with router where the application always rendered the tabs before anything else. The sign in modal is rendered on startup on the index page, but needs to be changed as it is quite abrupt and in the users face. A provider and hook have been made but don't have much functionality. Next step is to complete user authentication and test with predefined accounts.