Added a login button and conditional rendering.
If the user is not logged in, the index displays a login button. This is so if they close the login popup, it can be called again. Could be changed so that it is only callable. Added this to addLayout too, so when the user presses ready to go, it only shows the page if the websocket is connected. Need to fix the login button styling.