Moved db management into provider and optimised.
Since access to the db will be required in the settings page (deleting accounts, changing names etc), the db setup has been moved to authorisationProvider and provided as a hook. Now the system can call this hook to access the user and db. I have started work on the user-based settings, including account settings. A sign out button has been added and tested. Optimised the login modal by adding a new function "checks" which checks that the db works and grabs the first instance of the provided username.