Added password hashing.
For cybersecurity, the passwords stored in the sql db are now hashed. I wanted to use bcrypt but its not supported on expo, only crypto is. This is because expo is on the front end, whereas professional instances use the backend. Since this is just a demo, this hashing will do.