The layouts now save correctly, but duplicate.
The layouts now save by parsing in the old data and new data at once. This is an easier fix than trying to append data to the json which requires finding elements by index value, ensuring data is not overwritten etc etc. Not sure why but when saving the layouts, it duplicates the data per layout. I think this might be caused by a communication issue regarding variable names and data parsing. Need to talk to teammates to fix.