Library now loads layouts and layouts have titles.
The library now accesses the layouts json and displays the users layouts. Fixed an overwriting issue in addLayout by grabbing the other saved layouts and ensuring they remain when writing the new data. Each layout now requires a title, helping to identify it and display in the library. Still need to fix the duplication error and the library page needs to reload more dynamically, currently only reloading on user changes.