Layouts are now pressable and load layout page.
Layouts send across their title using link and params, loading the addLayout page. The addlayout page now listens to when users press the saved layout and updates the page title. Currently it only updates the title but it does get the ID's of each box, so can get their information. Need to figure out how to set the boxes correctly.