Fixed layout issues regarding keys and names.
Fixed the issue where layouts could be saved under the same name. In turn this (as well as setting the smallLayout keys) fixed the child props issue in the library by ensuring they had keys and they were unique. Tidied up addLayout a bit. Updated the readme with an alternative way of opening the app on Expo Go.