Deleting individual layouts works with bug.
Each layout can now be deleted when selected. Visuals have been improved when a layout is saved. There is a strange bug regarding "removeView" where it needs to be called. Occurs when a layout is deleted, then the user opens addLayout followed by library. Need to fix but not critical.