Added a .pgm to .png converter in comms_hub
Since the data will be parsed across as a .pgm file, it needs to be converted to .png and parsed over as base64 to be seen in the app. This script takes the .pgm, reformats it as a .png, outputs the base64 string, and sends it over to the hub.