Added the grid to the system running page.
Copied across the grid from the addLayout page, keeping the key functionality and removing the rest. It now shows the grid, the current title, and displays the box coordinates when selected. Need to fix the design and coordinates. Added extra changes to a few other pages, tidying functions and improving the return view.