Fixed issues with QR code and boxes on the grid.
The QR code displays when the websocket has qrcode data and this data would not clear, resulting in the qr code always being visible, even after the form resets. Now the qr code local variable (QRData) only updates once the save furniture button has been pressed, getting the new QR code. After fixing the issue with loading box coordinates from the library, it meant that adding a layout normally would result in no boxes appearing at the start. This was due to an incorrect condition only checking if the useEffect had been called. Now it checks that as well as if the selectedLayout is not empty, since it would return {}, counting as not empty. It now checks the key length within, rendering the correct functions only when needed.