systemRunning now moves the boxes by robot pos.
Using a use effect, system running now moves the boxes to a set position based on the robot's information. This was done in robot-testing.py but should work the same for the real robot provided that it sends the data in the correct format. It works for x, y, and rotation.