Fixed saving layout issues and added overwriting.
The biggest issue preventing the live simulation working was that the layouts would not overwrite just simply append with a new name. Before running the simulation, the layouts need to be saved to update for any changes. With this, the user would have to change the layout name, making saving redundant. Now the system checks if its a new or loaded layout and overwrites it accordingly. For the simulation, the grey boxes represent the target destinations instead of the "current layout", as an easy-to-understand visual for the user.