Added a file path input to updateJsonData.
Fixed the previous issue by adding an optional filepath input for updateJsonData. This way, the function can be called for direct edits which is quicker and is useful when the roomName could be changing (multiple smallRooms). If none is supplied, it defaults to roomPath which is used when the roomName is consistent i.e. when a room has been selected.